Rename to EnterpriseBasicTrialModal
